pyproject.toml
[project]
name = "netdevops-toolkit"
version = "0.1.0"
description = "NetDevOps parser toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "jinja2>=3.1.3",
  "jmespath>=1.0.1",
  "pyyaml>=6.0.1",
  "python-multipart>=0.0.9",
  "textfsm>=1.1.3",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[tool.uv]
package = false
